Circular dependencies eliminated
We could probably skip the reject list, at least for those that depend on themselves.

This closes #25

This also closes #16, because it's going to the root of the problem.

It should have solved #23... but it does not. Still, Test::Meta has higher score than META6, even if one depends on the other.

Still not solved #24. Will probably have to wait.